Update Get recipe by id endpoint

+++ CulinaryRecipesAPI/CulinaryRecipes.Domain/Recipe.cs
@@ -7,13 +7,13 @@ namespace CulinaryRecipes.Domain
 		public string Id { get; set; } = string.Empty;
 		public string Name { get; set; } = string.Empty;
 		public string Description { get; set; } = string.Empty;
-		public int PreparationTime { get; set; }
-		public int CookingTime { get; set; }
+		public long PreparationTime { get; set; }
+		public long CookingTime { get; set; }
 		public string SkillLevel { get; set; } = string.Empty;
 
 		private Recipe() {}
 
-		public static Recipe Create(string id, string name, string description, int preparationTime, int cookingTime,
+		public static Recipe Create(string id, string name, string description, long preparationTime, long cookingTime,
 			string skillLevel)
 		{
 			return new Recipe
